+++ nbgrader/nbextensions/static/assignment_list/assignment_list.js
@@ -9,6 +9,10 @@ define([
 ], function(Jupyter, $, utils, dialog) {
     "use strict";
 
+    var ajax = utils.ajax || $.ajax;
+    // Notebook v4.3.1 enabled xsrf so use notebooks ajax that includes the
+    // xsrf token in the header data
+
     var AssignmentList = function (released_selector, fetched_selector, submitted_selector, options) {
         this.released_selector = released_selector;
         this.fetched_selector = fetched_selector;
@@ -43,7 +47,7 @@ define([
             error : utils.log_ajax_error,
         };
         var url = utils.url_join_encode(this.base_url, 'assignments');
-        $.ajax(url, settings);
+        ajax(url, settings);
     };
 
 
@@ -95,7 +99,7 @@ define([
                     // to do the animation (borderSpacing).
                     $icon.animate({ borderSpacing: 90 }, {
                         step: function(now,fx) {
-                            $icon.css('transform','rotate(-' + now + 'deg)'); 
+                            $icon.css('transform','rotate(-' + now + 'deg)');
                         }
                     }, 250);
                 } else {
@@ -103,7 +107,7 @@ define([
                     // See comment above.
                     $icon.animate({ borderSpacing: 0 }, {
                         step: function(now,fx) {
-                            $icon.css('transform','rotate(-' + now + 'deg)'); 
+                            $icon.css('transform','rotate(-' + now + 'deg)');
                         }
                     }, 250);
                 }
@@ -226,7 +230,7 @@ define([
                     'assignments',
                     'fetch'
                 );
-                $.ajax(url, settings);
+                ajax(url, settings);
             });
 
         } else if (this.data.status == 'fetched') {
@@ -253,7 +257,7 @@ define([
                     'assignments',
                     'submit'
                 );
-                $.ajax(url, settings);
+                ajax(url, settings);
             });
         }
 
@@ -321,7 +325,7 @@ define([
                 'assignments',
                 'validate'
             );
-            $.ajax(url, settings);
+            ajax(url, settings);
         });
 
         return container;